Kw 585

+++ Daemons/OutboundEmail.cs
@@ -96,9 +96,7 @@ namespace Daemons
                             IEmailPackager packager = ObjectFactory.GetNamedInstance<IEmailPackager>(envelope.Handler);
                             if (configRepository.Get<bool>("ArchiveOutboundEmail"))
                             {
-                                EmailAddressDO outboundemailaddress =
-                                    subUow.EmailAddressRepository.GetOrCreateEmailAddress(
-                                        configRepository.Get("ArchiveEmailAddress"), "Outbound Archive");
+                                EmailAddressDO outboundemailaddress = subUow.EmailAddressRepository.GetOrCreateEmailAddress(configRepository.Get("ArchiveEmailAddress"), "Outbound Archive");
                                 envelope.Email.AddEmailRecipient(EmailParticipantType.Bcc, outboundemailaddress);
                             }
 
@@ -114,11 +112,15 @@ namespace Daemons
                                 }
                             }
 
+                            if (String.IsNullOrEmpty(envelope.Email.ReplyToAddress))
+                                envelope.Email.ReplyToAddress = configRepository.Get("replyToEmail", String.Empty);
+                            if (String.IsNullOrEmpty(envelope.Email.ReplyToName))
+                                envelope.Email.ReplyToName = configRepository.Get("replyToName", String.Empty);
+                            
                             packager.Send(envelope);
                             numSent++;
 
                             var email = envelope.Email;
-                                // subUow.EmailRepository.GetQuery().First(e => e.Id == envelope.Email.Id);
                             email.EmailStatus = EmailState.Dispatched;
                             subUow.SaveChanges();
 
@@ -160,7 +162,8 @@ namespace Daemons
             foreach (RecipientDO recipient in recipientList)
             {
                 UserDO user = uow.UserRepository.FindOne(e => e.EmailAddress.Address == recipient.EmailAddress.Address);
-                if (user != null && !user.TestAccount)
+
+                if (user != null && !uow.AspNetUserRolesRepository.UserHasRole(Roles.Admin, user.Id) && !user.TestAccount)
                 {
                     recipientsRemoved.Add(recipient.EmailAddress.Address);
                     uow.RecipientRepository.Remove(recipient);
